feat: typescript
ilp-plugin-xrp-asym-server and others should probably keep importing this using `require()` for now. Much of its "api" is exposed by overwriting various methods and properties, some of which are duplicated or change types which makes type checking troublesome.
