+++ .gitignore
@@ -1,1 +1,4 @@
 /node_modules/
+/src/**/*.js
+/src/**/*.js.map
+/src/**/*.d.ts
