+++ package.json
@@ -3,9 +3,13 @@
   "version": "3.4.9",
   "description": "Plugin that implements a mini ephemeral ledger",
   "main": "index.js",
+  "types": "src/index.d.js",
   "scripts": {
+    "build": "tsc --project .",
     "test": "mocha",
-    "lint": "standard"
+    "prepare": "npm run build",
+    "pretest": "npm run build",
+    "lint": "tslint --project ."
   },
   "repository": {
     "type": "git",
@@ -27,11 +31,12 @@
   },
   "homepage": "https://github.com/justmoon/ilp-plugin-mini-accounts#readme",
   "dependencies": {
+    "@types/node": "^10.5.1",
+    "@types/ws": "^5.1.2",
     "btp-packet": "^1.2.1",
-    "debug": "^3.1.0",
     "ilp-logger": "^1.0.1",
     "ilp-packet": "^2.2.0",
-    "ilp-plugin-btp": "^1.1.3",
+    "ilp-plugin-btp": "^1.1.14",
     "ilp-protocol-ildcp": "^1.0.0",
     "ilp-store-memory": "^1.0.0",
     "ilp-store-wrapper": "^1.0.0",
@@ -44,7 +49,10 @@
     "get-port": "^3.2.0",
     "mocha": "^5.0.0",
     "sinon": "^4.5.0",
-    "standard": "^11.0.1"
+    "ts-node": "^7.0.0",
+    "tslint": "^5.10.0",
+    "tslint-config-standard": "^7.1.0",
+    "typescript": "^2.9.2"
   },
   "engines": {
     "node": ">=8.3.0"
