+++ tslint.json
@@ -1,0 +1,8 @@
+{
+  "extends": "tslint-config-standard",
+  "linterOptions": {
+    "exclude": [
+      "test/**/*"
+    ]
+  }
+}
